\name{collexclean-package}
\alias{collexclean-package}
\alias{collexclean}
\docType{package}
\title{
\packageTitle{collexclean}
}
\description{
\packageDescription{collexclean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{collexclean}
\packageIndices{collexclean}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{collexclean}

Maintainer: Luis J. Villanueva
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
