\name{collexScrub-package}
\alias{collexScrub-package}
\alias{collexScrub}
\docType{package}
\title{
\packageTitle{collexScrub}
}
\description{
\packageDescription{collexScrub}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{collexScrub}
\packageIndices{collexScrub}
An overview of how to use the package, details to come.
}
\author{
\packageAuthor{collexScrub}

Maintainer: Luis J. Villanueva
}
\keyword{ package }
