\name{collexScrubberber-package}
\alias{collexScrubber-package}
\alias{collexScrubber}
\docType{package}
\title{
\packageTitle{collexScrubber}
}
\description{
\packageDescription{collexScrubber}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{collexScrubber}
\packageIndices{collexScrubber}
An overview of how to use the package, details to come.
}
\author{
\packageAuthor{collexScrubber}

Maintainer: Luis J. Villanueva
}
\keyword{ package }
