% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlTL2.R
\name{dlTL2}
\alias{dlTL2}
\title{Download the Taxonomic Literature data dump}
\usage{
dlTL2()
}
\value{
A dataframe of \code{authors}.
}
\description{
Function to download and load to a data.frame the 
list of authors in: 
Taxonomic Literature: A selective guide to botanical 
publications and collections with dates, commentaries
and types (Stafleu et al.).
}
\details{
http://www.sil.si.edu/DigitalCollections/tl-2/index.cfm
}
