% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms2dd.R
\name{dms2dd}
\alias{dms2dd}
\title{Convert lat/long from DMS to DD}
\usage{
dms2dd(deg = 0, min = 0, sec = 0, direction = "N", roundto = 5)
}
\arguments{
\item{deg}{Value of the degrees, \code{integer}}

\item{min}{Value of the minutes, \code{integer}}

\item{sec}{Value of the seconds, \code{float}}

\item{direction}{Which direction, \code{N, S, E, or W}}

\item{roundto}{How many decimal places to round to, default 5}
}
\description{
Function that converts a messy DMS to a DD format.
}
