% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_match_str.R
\name{find_match_str}
\alias{find_match_str}
\title{Find a string in a database of known strings}
\usage{
find_match_str(str_to_check, database, method = "osa", threshold = 10,
  no_cores = 2, year_limits = FALSE, country_limits = FALSE,
  database_strings = NA, str_to_check_col = NA)
}
\description{
Function to check a string that has problems due to 
incomplete transcription, OCR problems, illegible 
words, and other problems.
}
