% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sciname_paleo.R
\name{find_sciname_paleo}
\alias{find_sciname_paleo}
\title{Find the matches of a name in both the PaleoDB and WoRMS databases}
\usage{
find_sciname_paleo(input_name, sep_char = "|")
}
\arguments{
\item{input_name}{A scientific name to check against PaleoDB and WoRMS}

\item{sep_char}{Separating character for strings of \code{input_name}, the function will search only the last value. Default is pipe (|)}
}
\value{
A dataframe of species matched.
}
\description{
Find the matches of a name in both the PaleoDB and WoRMS databases
}
