% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCollectorsBotany.R
\name{findCollectorsBotany}
\alias{findCollectorsBotany}
\title{Search collectors in Harvard University Herbaria & Libraries}
\usage{
findCollectorsBotany(collector, year = NA, country = NA)
}
\arguments{
\item{collector}{Name of the collector}

\item{year}{Year when active}

\item{country}{Country where collector collected specimens}
}
\value{
A list of collectors.
}
\description{
Search collectors in Harvard University Herbaria & Libraries
}
\details{
https://kiki.huh.harvard.edu/databases/botanist_index.html
}
