% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aat.R
\name{query_aat}
\alias{query_aat}
\title{Query the Getty AAT SPARQL Service}
\usage{
query_aat(term = NA, query_type = NA, keywords_string = NA)
}
\arguments{
\item{term}{String to check against the AAT}

\item{query_type}{Which query to use, one of: 'direct', 'anylabel', 'fulltext', or 'widesearch'}

\item{keywords_string}{String or list to extract keywords from to try to get a better match (optional)}
}
\value{
A dataframe of \code{term} (the string provided),
\code{aat_term} (the matched term from AAT), \code{aat_id} (the id of the term, prefixed with 'aat:'), \code{aat_note} (Notes of the term), \code{aat_parents} (parent categories of the term).
}
\description{
Function to query the Getty Art & Architecture Thesaurus.
}
