% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_coords.R
\name{query_coords}
\alias{query_coords}
\title{Query a pair of lat/lon values against a spatial database}
\usage{
query_coords(coords = NA, database = "GADM", db_shp = "data",
  geom = FALSE)
}
\value{
A dataframe of matched geometries.
}
\description{
Query a pair of lat/lon values against a spatial database
}
